+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.0.2
+- Add roll tables in the Referee Dashboard for Myths and extra tables. (Sky/Weather, Person, Holding, Land, Beast)
+- Fix spelling for Fatigued and Armour.
+- Correct typos in some compendium entries.
+
 # v1.0.1
 - Fix bonus dice being ignore when attacking
 - Fix empty button on taking damage